added copykey job and tasks
