+++ lib/task-data/base-tasks/copy-key.js
@@ -1,0 +1,13 @@
+// Copyright 2016, EMC, Inc.
+
+'use strict';
+
+
+module.exports = {
+    friendlyName: 'Copy Key',
+    injectableName: 'Task.Base.Ssh.CopyKey',
+    runJob: 'Job.CopyKey',
+    requiredOptions: [],
+    requiredProperties: {},
+    properties: {}
+};
