+++ lib/task-data/tasks/copy-publickey.js
@@ -1,0 +1,11 @@
+// Copyright 2016, EMC, Inc.
+
+'use strict';
+
+module.exports = {
+    friendlyName: 'Copy Public Key',
+    injectableName: 'Task.Ssh.CopyKey',
+    implementsTask: 'Task.Base.Ssh.CopyKey',
+    options: {},
+    properties: {}
+};
