+++ lib/utils/job-utils/command-util.js
@@ -12,6 +12,7 @@ di.annotate(commandUtilFactory, new di.Inject(
     'Assert',
     'Promise',
     'Services.Waterline',
+    'Services.Encryption',
     '_'
 ));
 
@@ -21,6 +22,7 @@ function commandUtilFactory(
     assert,
     Promise,
     waterline,
+    cryptService,
     _
 ) {
     var logger = Logger.initialize(commandUtilFactory);
@@ -103,6 +105,67 @@ function commandUtilFactory(
             }
         });
      };
+
+    CommandUtil.prototype.sshExec = function(cmdObj, sshSettings, sshClient) {
+        var self = this;
+        return new Promise(function(resolve, reject) {
+            if(cmdObj.timeout) {
+                setTimeout(function() {
+                    var seconds = cmdObj.timeout / 1000;
+                    reject(new Error('The remote operation timed out after '+
+                                 seconds + ' seconds'));
+                }, cmdObj.timeout);
+            }
+            var ssh = sshClient;
+            ssh.on('ready', function() {
+                ssh.exec(cmdObj.cmd, function(err, stream) {
+                    if (err) { reject(err); }
+                    stream.on('close', function(code) {
+                        cmdObj.exitCode = code;
+                        ssh.end();
+                    }).on('error', function(err) {
+                        reject(err);
+                    }).on('data', function(data) {
+                        cmdObj.stdout = ( cmdObj.stdout || '' ) + data.toString();
+                    }).stderr.on('data', function(data) {
+                        cmdObj.stderr = ( cmdObj.stderr || '' ) + data.toString();
+                    });
+                });
+            })
+            .on('error', function(err) {
+                logger.error('ssh error', {
+                    error: err,
+                    host: sshSettings.host,
+                    task: cmdObj,
+                    level: err.level,
+                    description: err.description
+                });
+                reject(err);
+            })
+            .on('close', function(hasErr) {
+
+                if (hasErr || (cmdObj.exitCode &&
+                    !_.contains(cmdObj.acceptedResponseCodes, cmdObj.exitCode))) {
+                    logger.error("Failure running remote command", {task:cmdObj});
+
+                    reject(new Error(
+                            "Encountered a failure running "+cmdObj.cmd+
+                            "on remote host"+ sshSettings.host +". NodeId: "+
+                            self.nodeId
+                    ));
+                } else {
+                    resolve(cmdObj);
+                }
+            });
+            ssh.connect({
+                host: sshSettings.host,
+                port: sshSettings.port || 22,
+                username: sshSettings.user,
+                password: cryptService.decrypt(sshSettings.password),
+                privateKey: cryptService.decrypt(sshSettings.privateKey)
+            });
+        });
+    };
 
     CommandUtil.prototype.buildCommands = function(commands) {
         return _.map(_.flatten([commands]), function(cmd) {
